[wasm][debugger][tests] Move helper methods to base class
