+++ sdks/wasm/DebuggerTestSuite/Tests.cs
@@ -1,13 +1,6 @@
 using System;
 using System.Linq;
 using System.Threading.Tasks;
-
-using System.Net.WebSockets;
-using System.Threading;
-using System.IO;
-using System.Text;
-using System.Collections.Generic;
-
 using Newtonsoft.Json.Linq;
 using Xunit;
 using WebAssembly.Net.Debugging;
@@ -16,39 +9,6 @@ namespace DebuggerTests
 {
 
 	public class SourceList : DebuggerTestBase {
-		DebugTestContext ctx;
-		Dictionary<string, string> dicScriptsIdToUrl;
-		Dictionary<string, string> dicFileToUrl;
-		Dictionary<string, string> SubscribeToScripts (Inspector insp) {
-			dicScriptsIdToUrl = new Dictionary<string, string> ();
-			dicFileToUrl = new Dictionary<string, string>();
-			insp.On("Debugger.scriptParsed", async (args, c) => {
-				var script_id = args? ["scriptId"]?.Value<string> ();
-				var url = args["url"]?.Value<string> ();
-				if (script_id.StartsWith("dotnet://"))
-				{
-					var dbgUrl = args["dotNetUrl"]?.Value<string>();
-					var arrStr = dbgUrl.Split("/");
-					dbgUrl = arrStr[0] + "/" + arrStr[1] + "/" + arrStr[2] + "/" + arrStr[arrStr.Length - 1];
-					dicScriptsIdToUrl[script_id] = dbgUrl;
-					dicFileToUrl[dbgUrl] = args["url"]?.Value<string>();
-				} else if (!String.IsNullOrEmpty (url)) {
-					dicFileToUrl[new Uri (url).AbsolutePath] = url;
-				}
-				await Task.FromResult (0);
-			});
-			return dicScriptsIdToUrl;
-		}
-
-		void CheckLocation (string script_loc, int line, int column, Dictionary<string, string> scripts, JToken location)
-		{
-			var loc_str = $"{ scripts[location["scriptId"].Value<string>()] }"
-							+ $"#{ location ["lineNumber"].Value<int> () }"
-							+ $"#{ location ["columnNumber"].Value<int> () }";
-
-			var expected_loc_str = $"{script_loc}#{line}#{column}";
-			Assert.Equal (expected_loc_str, loc_str);
-		}
 
 		[Fact]
 		public async Task CheckThatAllSourcesAreSent () {
@@ -204,143 +164,6 @@ namespace DebuggerTests
 				Assert.Equal (dicFileToUrl["/debugger-driver.html"], ex_json ["exceptionDetails"]? ["url"]? .Value<string> ());
 			});
 
-		}
-
-		void CheckNumber (JToken locals, string name, int value) {
-			foreach (var l in locals) {
-				if (name != l["name"]?.Value<string> ())
-					continue;
-				var val = l["value"];
-				Assert.Equal ("number", val ["type"]?.Value<string> ());
-				Assert.Equal (value, val["value"]?.Value <int> ());
-				return;
-			}
-			Assert.True(false, $"Could not find variable '{name}'");
-		}
-
-		void CheckString (JToken locals, string name, string value) {
-			foreach (var l in locals) {
-				if (name != l["name"]?.Value<string> ())
-					continue;
-				var val = l["value"];
-				if (value == null) {
-						Assert.Equal ("object", val ["type"]?.Value<string> ());
-						Assert.Equal ("null", val["subtype"]?.Value<string> ());
-				} else {
-						Assert.Equal ("string", val ["type"]?.Value<string> ());
-						Assert.Equal (value, val["value"]?.Value <string> ());
-				}
-				return;
-			}
-			Assert.True(false, $"Could not find variable '{name}'");
-		}
-
-		JToken CheckSymbol (JToken locals, string name, string value)
-		{
-			var l = GetAndAssertObjectWithName (locals, name);
-			var val = l["value"];
-			Assert.Equal ("symbol", val ["type"]?.Value<string> ());
-			Assert.Equal (value, val ["value"]?.Value<string> ());
-			return l;
-		}
-
-		JToken CheckObject (JToken locals, string name, string class_name, string subtype=null, bool is_null=false) {
-			var l = GetAndAssertObjectWithName (locals, name);
-			var val = l["value"];
-			Assert.Equal ("object", val ["type"]?.Value<string> ());
-			Assert.True (val ["isValueType"] == null || !val ["isValueType"].Value<bool> ());
-			Assert.Equal (class_name, val ["className"]?.Value<string> ());
-
-			var has_null_subtype = val ["subtype"] != null && val ["subtype"]?.Value<string> () == "null";
-			Assert.Equal (is_null, has_null_subtype);
-			if (subtype != null)
-				Assert.Equal (subtype, val ["subtype"]?.Value<string> ());
-
-			return l;
-		}
-
-		async Task CheckDateTime (JToken locals, string name, DateTime expected)
-		{
-			var obj = GetAndAssertObjectWithName(locals, name);
-			await CheckDateTimeValue (obj ["value"], expected);
-		}
-
-		async Task CheckDateTimeValue (JToken value, DateTime expected)
-		{
-			AssertEqual ("System.DateTime", value ["className"]?.Value<string> (), "className");
-			AssertEqual (expected.ToString (), value ["description"]?.Value<string> (), "className");
-
-			var members = await GetProperties (value ["objectId"]?.Value<string> ());
-
-			// not checking everything
-			CheckNumber (members, "Year", expected.Year);
-			CheckNumber (members, "Month", expected.Month);
-			CheckNumber (members, "Day", expected.Day);
-			CheckNumber (members, "Hour", expected.Hour);
-			CheckNumber (members, "Minute", expected.Minute);
-			CheckNumber (members, "Second", expected.Second);
-
-			// FIXME: check some float properties too
-		}
-
-		JToken CheckBool (JToken locals, string name, bool expected)
-		{
-			var l = GetAndAssertObjectWithName (locals, name);
-			var val = l["value"];
-			Assert.Equal ("boolean", val ["type"]?.Value<string> ());
-			if (val ["value"] == null)
-				Assert.True (false, "expected bool value not found for variable named {name}");
-			Assert.Equal (expected, val ["value"]?.Value<bool> ());
-
-			return l;
-		}
-
-		void CheckContentValue (JToken token, string value) {
-			var val = token["value"].Value<string> ();
-			Assert.Equal (value, val);
-		}
-
-		JToken CheckValueType (JToken locals, string name, string class_name) {
-			var l = GetAndAssertObjectWithName (locals, name);
-			var val = l["value"];
-			Assert.Equal ("object", val ["type"]?.Value<string> ());
-			Assert.True (val ["isValueType"] != null && val ["isValueType"].Value<bool> ());
-			Assert.Equal (class_name, val ["className"]?.Value<string> ());
-			return l;
-		}
-
-		JToken CheckEnum (JToken locals, string name, string class_name, string descr) {
-			var l = GetAndAssertObjectWithName (locals, name);
-			var val = l["value"];
-			Assert.Equal ("object", val ["type"]?.Value<string> ());
-			Assert.True (val ["isEnum"] != null && val ["isEnum"].Value<bool> ());
-			Assert.Equal (class_name, val ["className"]?.Value<string> ());
-			Assert.Equal (descr, val ["description"]?.Value<string> ());
-			return l;
-		}
-
-		void CheckArray (JToken locals, string name, string class_name) {
-			foreach (var l in locals) {
-				if (name != l["name"]?.Value<string> ())
-					continue;
-
-				var val = l["value"];
-				Assert.Equal ("object", val ["type"]?.Value<string> ());
-				Assert.Equal ("array", val ["subtype"]?.Value<string> ());
-				Assert.Equal (class_name, val ["className"]?.Value<string> ());
-
-				//FIXME: elements?
-				return;
-			}
-			Assert.True(false, $"Could not find variable '{name}'");
-		}
-
-		JToken GetAndAssertObjectWithName (JToken obj, string name)
-		{
-			var l = obj.FirstOrDefault (jt => jt ["name"]?.Value<string> () == name);
-			if (l == null)
-				Assert.True (false, $"Could not find variable '{name}'");
-			return l;
 		}
 
 		[Theory]
@@ -2768,34 +2591,6 @@ namespace DebuggerTests
 					// await CheckDateTimeValue (evaluate, new DateTime (2000, 5, 4, 3, 2, 1));
 				});
 
-		async Task<Result> SendCommand (string method, JObject args) {
-			var res = await ctx.cli.SendCommand (method, args, ctx.token);
-			if (!res.IsOk) {
-				Console.WriteLine ($"Failed to run command {method} with args: {args?.ToString ()}\nresult: {res.Error.ToString ()}");
-				Assert.True (false, $"SendCommand for {method} failed with {res.Error.ToString ()}");
-			}
-			return res;
-		}
-
-		async Task<Result> Evaluate (string expression) {
-			return await SendCommand ("Runtime.evaluate", JObject.FromObject (new { expression = expression }));
-		}
-
-		void AssertLocation (JObject args, string methodName) {
-			Assert.Equal (methodName, args ["callFrames"]?[0]?["functionName"]?.Value<string> ());
-		}
-
-		// Place a breakpoint in the given method and run until its hit
-		// Return the Debugger.paused data
-		async Task<JObject> RunUntil (string methodName) {
-			await SetBreakpointInMethod ("debugger-test", "DebuggerTest", methodName);
-			// This will run all the tests until it hits the bp
-			await Evaluate ("window.setTimeout(function() { invoke_run_all (); }, 1);");
-			var wait_res = await ctx.insp.WaitFor (Inspector.PAUSE);
-			AssertLocation (wait_res, "locals_inner");
-			return wait_res;
-		}
-
 		[Fact]
 		public async Task InspectLocals () {
 			var insp = new Inspector ();
@@ -2879,425 +2674,8 @@ namespace DebuggerTests
 
 				});
 
-		async Task<JObject> StepAndCheck (StepKind kind, string script_loc, int line, int column, string function_name,
-							Func<JObject, Task> wait_for_event_fn = null, Action<JToken> locals_fn = null, int times=1)
-		{
-			for (int i = 0; i < times - 1; i ++) {
-				await SendCommandAndCheck (null, $"Debugger.step{kind.ToString ()}", null, -1, -1, null);
-			}
-
-			// Check for method/line etc only at the last step
-			return await SendCommandAndCheck (
-						null, $"Debugger.step{kind.ToString ()}", script_loc, line, column, function_name,
-						wait_for_event_fn: wait_for_event_fn,
-						locals_fn: locals_fn);
-		}
-
-		async Task<JObject> EvaluateAndCheck (string expression, string script_loc, int line, int column, string function_name,
-								Func<JObject, Task> wait_for_event_fn = null, Action<JToken> locals_fn = null)
-			=> await SendCommandAndCheck (
-						JObject.FromObject (new { expression = expression }),
-						"Runtime.evaluate", script_loc, line, column, function_name,
-						wait_for_event_fn: wait_for_event_fn,
-						locals_fn: locals_fn);
-
-		async Task<JObject> SendCommandAndCheck (JObject args, string method, string script_loc, int line, int column, string function_name,
-								Func<JObject, Task> wait_for_event_fn = null, Action<JToken> locals_fn = null, string waitForEvent = Inspector.PAUSE)
-		{
-			var res = await ctx.cli.SendCommand (method, args, ctx.token);
-			if (!res.IsOk) {
-				Console.WriteLine ($"Failed to run command {method} with args: {args?.ToString ()}\nresult: {res.Error.ToString ()}");
-				Assert.True (false, $"SendCommand for {method} failed with {res.Error.ToString ()}");
-			}
-
-			var wait_res = await ctx.insp.WaitFor(waitForEvent);
-
-			if (function_name != null)
-				Assert.Equal (function_name, wait_res ["callFrames"]?[0]?["functionName"]?.Value<string> ());
-
-			if (script_loc != null)
-				CheckLocation (script_loc, line, column, ctx.scripts, wait_res ["callFrames"][0]["location"]);
-
-			if (wait_for_event_fn != null)
-				await wait_for_event_fn (wait_res);
-
-			if (locals_fn != null) {
-				var locals = await GetProperties (wait_res ["callFrames"][0]["callFrameId"].Value<string> ());
-				locals_fn (locals);
-			}
-
-			return wait_res;
-		}
-
-		async Task CheckDelegate (JToken locals, string name, string className, string target)
-		{
-			var l = GetAndAssertObjectWithName (locals, name);
-			var val = l["value"];
-
-			await CheckDelegate (l, TDelegate (className, target), name);
-		}
-
-		async Task CheckDelegate (JToken actual_val, JToken exp_val, string label)
-		{
-			AssertEqual ("object", actual_val["type"]?.Value<string>(), $"{label}-type");
-			AssertEqual (exp_val ["className"]?.Value<string> (), actual_val ["className"]?.Value<string>(), $"{label}-className");
-
-			var actual_target = actual_val["description"]?.Value<string>();
-			Assert.True(actual_target != null, $"${label}-description");
-			var exp_target = exp_val["target"].Value<string>();
-
-			CheckDelegateTarget (actual_target, exp_target);
-
-			var del_props = await GetProperties(actual_val["objectId"]?.Value<string>());
-			AssertEqual (1, del_props.Count(), $"${label}-delegate-properties-count");
-
-			var obj = del_props.Where (jt => jt ["name"]?.Value<string> () == "Target").FirstOrDefault ();
-			Assert.True (obj != null, $"[{label}] Property named 'Target' found found in delegate properties");
-
-			AssertEqual("symbol", obj ["value"]?["type"]?.Value<string>(), $"{label}#Target#type");
-			CheckDelegateTarget(obj ["value"]?["value"]?.Value<string>(), exp_target);
-
-			return;
-
-			void CheckDelegateTarget(string actual_target, string exp_target)
-			{
-				var parts = exp_target.Split(new char[] { '|' }, StringSplitOptions.RemoveEmptyEntries);
-				if (parts.Length == 1) {
-					// not a generated method
-					AssertEqual(exp_target, actual_target, $"{label}-description");
-				} else {
-					bool prefix = actual_target.StartsWith(parts[0], StringComparison.Ordinal);
-					Assert.True(prefix, $"{label}-description, Expected target to start with '{parts[0]}'. Actual: '{actual_target}'");
-
-					var remaining = actual_target.Substring(parts[0].Length);
-					bool suffix = remaining.EndsWith(parts[1], StringComparison.Ordinal);
-					Assert.True(prefix, $"{label}-description, Expected target to end with '{parts[1]}'. Actual: '{remaining}'");
-				}
-			}
-		}
-
-		async Task CheckCustomType (JToken actual_val, JToken exp_val, string label)
-		{
-			var ctype = exp_val["__custom_type"].Value<string>();
-			switch (ctype) {
-				case "delegate":
-					await CheckDelegate (actual_val, exp_val, label);
-					break;
-
-				case "pointer": {
-					AssertEqual ("symbol", actual_val ["type"]?.Value<string>(), $"{label}-type");
-
-					if (exp_val ["is_null"]?.Value<bool>() == false) {
-						var exp_prefix = $"({exp_val ["type_name"]?.Value<string>()})";
-						AssertStartsWith (exp_prefix, actual_val ["value"]?.Value<string> (), $"{label}-type_name");
-						AssertStartsWith (exp_prefix, actual_val ["description"]?.Value<string> (), $"{label}-description");
-					} else {
-						var exp_prefix = $"({exp_val ["type_name"]?.Value<string>()}) 0";
-						AssertEqual (exp_prefix, actual_val ["value"]?.Value<string> (), $"{label}-type_name");
-						AssertEqual (exp_prefix, actual_val ["description"]?.Value<string> (), $"{label}-description");
-					}
-					break;
-				}
-
-				case "ignore_me":
-					// nothing to check ;)
-					break;
-
-				default:
-					throw new ArgumentException($"{ctype} not supported");
-			}
-		}
-
-		async Task CheckProps (JToken actual, object exp_o, string label, int num_fields=-1)
-		{
-			if (exp_o.GetType ().IsArray || exp_o is JArray) {
-				if (! (actual is JArray actual_arr)) {
-					Assert.True (false, $"[{label}] Expected to get an array here but got {actual}");
-					return;
-				}
-
-				var exp_v_arr = JArray.FromObject (exp_o);
-				AssertEqual (exp_v_arr.Count, actual_arr.Count (), $"{label}-count");
-
-				for (int i = 0; i < exp_v_arr.Count; i ++) {
-					var exp_i = exp_v_arr [i];
-					var act_i = actual_arr [i];
-
-					AssertEqual (i.ToString (), act_i ["name"]?.Value<string> (), $"{label}-[{i}].name");
-
-					await CheckValue (act_i["value"], exp_i, $"{label}-{i}th value");
-				}
-
-				return;
-			}
-
-			// Not an array
-			var exp = exp_o as JObject;
-			if (exp == null)
-				exp = JObject.FromObject(exp_o);
-
-			num_fields = num_fields < 0 ? exp.Values<JToken>().Count() : num_fields;
-			Assert.True(num_fields == actual.Count(), $"[{label}] Number of fields don't match, Expected: {num_fields}, Actual: {actual.Count()}");
-
-			foreach (var kvp in exp) {
-				var exp_name = kvp.Key;
-				var exp_val = kvp.Value;
-
-				var actual_obj = actual.FirstOrDefault(jt => jt["name"]?.Value<string>() == exp_name);
-				if (actual_obj == null) {
-					Console.WriteLine($"actual: {actual}, exp_name: {exp_name}, exp_val: {exp_val}");
-					Assert.True(actual_obj != null, $"[{label}] Could not find property named '{exp_name}'");
-				}
-
-				var actual_val = actual_obj["value"];
-				Assert.True(actual_obj != null, $"[{label}] not value found for property named '{exp_name}'");
-
-				if (exp_val.Type == JTokenType.Array) {
-					var actual_props = await GetProperties(actual_val["objectId"]?.Value<string>());
-					await CheckProps (actual_props, exp_val, $"{label}-{exp_name}");
-				} else {
-					await CheckValue (actual_val, exp_val, $"{label}#{exp_name}");
-				}
-			}
-		}
-
-		async Task CheckValue (JToken actual_val, JToken exp_val, string label)
-		{
-			if (exp_val ["__custom_type"] != null) {
-				await CheckCustomType (actual_val, exp_val, label);
-				return;
-			}
-
-			if (exp_val ["type"] == null && actual_val ["objectId"] != null) {
-				var new_val = await GetProperties (actual_val ["objectId"].Value<string> ());
-				await CheckProps (new_val, exp_val, $"{label}-{actual_val["objectId"]?.Value<string>()}");
-				return;
-			}
-
-			foreach (var jp in exp_val.Values<JProperty> ()) {
-				if (jp.Value.Type == JTokenType.Object) {
-					var new_val = await GetProperties (actual_val ["objectId"].Value<string> ());
-					await CheckProps (new_val, jp.Value, $"{label}-{actual_val["objectId"]?.Value<string>()}");
-
-					continue;
-				}
-
-				var exp_val_str = jp.Value.Value<string> ();
-				bool null_or_empty_exp_val = String.IsNullOrEmpty (exp_val_str);
-
-				var actual_field_val = actual_val.Values<JProperty> ().FirstOrDefault (a_jp => a_jp.Name == jp.Name);
-				var actual_field_val_str = actual_field_val?.Value?.Value<string> ();
-				if (null_or_empty_exp_val && String.IsNullOrEmpty (actual_field_val_str))
-					continue;
-
-				Assert.True (actual_field_val != null, $"[{label}] Could not find value field named {jp.Name}");
-
-				Assert.True (exp_val_str == actual_field_val_str,
-						$"[{label}] Value for json property named {jp.Name} didn't match.\n" +
-						$"Expected: {jp.Value.Value<string> ()}\n" +
-						$"Actual:   {actual_field_val.Value.Value<string> ()}");
-			}
-		}
-
-		async Task<JToken> GetLocalsForFrame (JToken frame, string script_loc, int line, int column, string function_name)
-		{
-			CheckLocation (script_loc, line, column, ctx.scripts, frame ["location"]);
-			Assert.Equal (function_name, frame ["functionName"].Value<string> ());
-
-			return await GetProperties (frame ["callFrameId"].Value<string> ());
-		}
-
-		async Task<JToken> GetObjectOnFrame (JToken frame, string name)
-		{
-			var locals = await GetProperties (frame ["callFrameId"].Value<string> ());
-			return await GetObjectOnLocals (locals, name);
-		}
-
-		// Find an object with @name, *fetch* the object, and check against @o
-		async Task<JToken> CompareObjectPropertiesFor (JToken locals, string name, object o, string label = null, int num_fields = -1)
-		{
-			if (label == null)
-				label = name;
-			var props = await GetObjectOnLocals (locals, name);
-			try {
-				if (o != null)
-					await CheckProps (props, o, label, num_fields);
-				return props;
-			} catch {
-				throw;
-			}
-		}
-
-		async Task<JToken> GetObjectOnLocals (JToken locals, string name)
-		{
-			var obj = GetAndAssertObjectWithName (locals, name);
-			var objectId = obj ["value"]["objectId"]?.Value<string> ();
-			Assert.True (!String.IsNullOrEmpty (objectId), $"No objectId found for {name}");
-
-			return await GetProperties (objectId);
-		}
-
-		/* @fn_args is for use with `Runtime.callFunctionOn` only */
-		async Task<JToken> GetProperties (string id, JToken fn_args = null)
-		{
-			if (ctx.UseCallFunctionOnBeforeGetProperties && !id.StartsWith ("dotnet:scope:")) {
-				var fn_decl = "function () { return this; }";
-				var cfo_args = JObject.FromObject (new {
-					functionDeclaration = fn_decl,
-					objectId = id
-				});
-				if (fn_args != null)
-					cfo_args ["arguments"] = fn_args;
-
-				var result = await ctx.cli.SendCommand ("Runtime.callFunctionOn", cfo_args, ctx.token);
-				AssertEqual (true, result.IsOk, $"Runtime.getProperties failed for {cfo_args.ToString ()}, with Result: {result}");
-				id = result.Value ["result"]?["objectId"]?.Value<string> ();
-			}
-
-			var get_prop_req = JObject.FromObject (new {
-				objectId = id
-			});
-
-			var frame_props = await ctx.cli.SendCommand ("Runtime.getProperties", get_prop_req, ctx.token);
-			if (!frame_props.IsOk)
-				Assert.True (false, $"Runtime.getProperties failed for {get_prop_req.ToString ()}, with Result: {frame_props}");
-
-			var locals = frame_props.Value ["result"];
-			// FIXME: Should be done when generating the list in library_mono.js, but not sure yet
-			//        whether to remove it, and how to do it correctly.
-			if (locals is JArray) {
-				foreach (var p in locals) {
-					if (p ["name"]?.Value<string> () == "length" && p ["enumerable"]?.Value<bool> () != true) {
-						p.Remove ();
-						break;
-					}
-				}
-			}
-
-			return locals;
-		}
-
-		async Task<JToken> EvaluateOnCallFrame (string id, string expression)
-		{
-			var evaluate_req = JObject.FromObject (new {
-				callFrameId = id,
-				expression = expression
-			});
-
-			var frame_evaluate = await ctx.cli.SendCommand ("Debugger.evaluateOnCallFrame", evaluate_req, ctx.token);
-			if (!frame_evaluate.IsOk)
-				Assert.True (false, $"Debugger.evaluateOnCallFrame failed for {evaluate_req.ToString ()}, with Result: {frame_evaluate}");
-
-			var evaluate_result = frame_evaluate.Value ["result"];
-			return evaluate_result;
-		}
-
-		async Task<Result> SetBreakpoint (string url_key, int line, int column, bool expect_ok=true)
-		{
-			var bp1_req = JObject.FromObject(new {
-				lineNumber = line,
-				columnNumber = column,
-				url = dicFileToUrl[url_key],
-			});
-
-			var bp1_res = await ctx.cli.SendCommand ("Debugger.setBreakpointByUrl", bp1_req, ctx.token);
-			Assert.True (expect_ok ? bp1_res.IsOk : bp1_res.IsErr);
-
-			return bp1_res;
-		}
-
-		async Task<Result> SetBreakpointInMethod (string assembly, string type, string method) {
-			var req = JObject.FromObject (new { assemblyName = assembly, typeName = type, methodName = method });
-
-			// Protocol extension
-			var res = await ctx.cli.SendCommand ("Dotnet-test.setBreakpointByMethod", req, ctx.token);
-			Assert.True (res.IsOk);
-
-			return res;
-		}
-
-		void AssertEqual (object expected, object actual, string label)
-			=> Assert.True (expected?.Equals (actual),
-						$"[{label}]\n" +
-						$"Expected: {expected?.ToString()}\n" +
-						$"Actual:   {actual?.ToString()}\n");
-
-		void AssertStartsWith (string expected, string actual, string label)
-			=> Assert.True(actual?.StartsWith (expected), $"[{label}] Does not start with the expected string\nExpected: {expected}\nActual:  {actual}");
-
-		//FIXME: um maybe we don't need to convert jobject right here!
-		static JObject TString (string value) =>
-			value == null
-				? TObject ("string", is_null: true)
-				: JObject.FromObject (new { type = "string", value = @value, description = @value });
-
-		static JObject TNumber (int value) =>
-			JObject.FromObject (new { type = "number", value = @value.ToString (), description = value.ToString () });
-
-		static JObject TValueType (string className, string description = null, object members = null) =>
-			JObject.FromObject (new { type = "object", isValueType = true, className = className, description = description ?? className });
-
-		static JObject TEnum (string className, string descr, object members = null) =>
-			JObject.FromObject (new { type = "object", isEnum = true, className = className, description = descr });
-
-		static JObject TObject (string className, string description = null, bool is_null = false) =>
-			is_null
-				? JObject.FromObject (new { type = "object", className = className, description = description ?? className, subtype = is_null ? "null" : null })
-				: JObject.FromObject (new { type = "object", className = className, description = description ?? className });
-
-		static JObject TArray (string className, int length = 0)
-			=> JObject.FromObject (new { type = "object", className = className, description = $"{className}({length})", subtype = "array" });
-
-		static JObject TBool (bool value)
-			=> JObject.FromObject (new { type = "boolean", value = @value, description = @value ? "true" : "false" });
-
-		static JObject TSymbol (string value)
-			=> JObject.FromObject (new { type = "symbol", value = @value, description = @value });
-
-		/*
-			For target names with generated method names like
-				`void <ActionTSignatureTest>b__11_0 (Math.GenericStruct<int[]>)`
-
-			.. pass target "as `target: "void <ActionTSignatureTest>|(Math.GenericStruct<int[]>)"`
-		*/
-		static JObject TDelegate(string className, string target)
-			=> JObject.FromObject(new {
-				__custom_type = "delegate",
-				className = className,
-				target = target
-			});
-
-		static JObject TPointer (string type_name, bool is_null = false)
-			=> JObject.FromObject (new { __custom_type = "pointer", type_name = type_name, is_null = is_null });
-
-		static JObject TIgnore ()
-			=> JObject.FromObject (new { __custom_type = "ignore_me" });
 		//
 		//TODO add tests covering basic stepping behavior as step in/out/over
 	}
 
-	class DebugTestContext
-	{
-		public InspectorClient cli;
-		public Inspector insp;
-		public CancellationToken token;
-		public Dictionary<string, string> scripts;
-
-		public bool UseCallFunctionOnBeforeGetProperties;
-
-		public DebugTestContext (InspectorClient cli, Inspector insp, CancellationToken token, Dictionary<string, string> scripts)
-		{
-				this.cli = cli;
-				this.insp = insp;
-				this.token = token;
-				this.scripts = scripts;
-		}
-	}
-
-	enum StepKind
-	{
-		Into,
-		Over,
-		Out
-	}
 }
